update-7/24-2

+++ database/migrations/2025_07_03_093421_create_inventory_transactions_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
             $table->boolean('transaction_type');
             $table->boolean('item_type');
             $table->integer('status')->default(0);
-            $table->integer('units')->nullable()->default(null);
+            $table->decimal('units')->nullable()->default(null);
             $table->foreignIdFor(Tea::class)->nullable()->constrained()->onDelete('set null');
             $table->foreignIdFor(Material::class)->nullable()->constrained()->onDelete('set null');
             $table->foreignIdFor(ProductionMaterial::class)->nullable()->constrained()->onDelete('set null');
